reworked DirtyDeeds state
