+++ src/commonMain/kotlin/character/classes/rogue/Rogue.kt
@@ -59,6 +59,7 @@ class Rogue(talents: Map<String, Talent>, spec: Spec) : Class(talents, spec) {
             character.classes.rogue.talents.BladeFlurry.name -> character.classes.rogue.talents.BladeFlurry(ranks)
             character.classes.rogue.talents.Preparation.name -> character.classes.rogue.talents.Preparation(ranks)
             character.classes.rogue.talents.ColdBlood.name -> character.classes.rogue.talents.ColdBlood(ranks)
+            character.classes.rogue.talents.DirtyDeeds.name -> character.classes.rogue.talents.DirtyDeeds(ranks)
             CombatPotency.name -> CombatPotency(ranks)
             Opportunity.name -> Opportunity(ranks)
             Aggression.name -> Aggression(ranks)
@@ -87,7 +88,6 @@ class Rogue(talents: Map<String, Talent>, spec: Spec) : Class(talents, spec) {
             Deflection.name -> Deflection(ranks)
             ImprovedSprint.name -> ImprovedSprint(ranks)
             SealFate.name -> SealFate(ranks)
-            DirtyDeeds.name -> DirtyDeeds(ranks)
             Initiative.name -> Initiative(ranks)
             SerratedBlades.name -> SerratedBlades(ranks)
             DirtyTricks.name -> DirtyTricks(ranks)
